+++ backend/data_validation/config/constants.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class RapidReact:
+    AUTO_TAXI_POINT_VALUE = 2
+
+    AUTO_UPPER_HUB_POINT_VALUE = 4
+    AUTO_LOWER_HUB_POINT_VALUE = 2
+
+    TELEOP_UPPER_HUB_POINT_VALUE = 2
+    TELEOP_LOWER_HUB_POINT_VALUE = 1
